Move GetAllReigsteredQueries to another file and add tests
